// src/types/user/salary/SalaryComponentCode.ts
export type SalaryComponentCode =
  | 'BASIC_GROSS_SALARY'
  | 'SOCIAL_INSURANCE_SALARY'
  | 'INSURANCE_BALANCE_ALLOWANCE'
  | 'JOB_ALLOWANCE'
  | 'PARKING_ALLOWANCE'
  | 'MANAGEMENT_ALLOWANCE'
  | 'MEAL_ALLOWANCE'
  | 'PHONE_ALLOWANCE'
  | 'FUEL_ALLOWANCE'
  | 'CLIENT_ENTERTAINMENT_ALLOWANCE'
  | 'TARGET_ALLOWANCE';